several plugins updated; Fixes for WXZY Quat ordering; more planning

+++ egg_plugins/src/GRN/grnitemlist.cpp
@@ -35,7 +35,7 @@ void ItemList::load(cGrannyStream * file)
 void ItemList::loadList(dword peers,cGrannyStream * file)
 {
 //  assert(file);
-	for (int i=0;i<peers;)
+  for (int i=0;i<(int)peers;)
 	{
 		dword chunk=file->readDword();
 		dword offset=file->readDword();
